refactor: rename service to ValdemortConfig
